# Increase speed of bicubic spline interpolation