By computing the interval that bounds the point in advance, the
repeated spline evaluation can be performed much faster.

Some timing for 1 million bicubic spline interpolations:

| Before | After | Speedup |
|--|--|--|
| 3.09155s | 1.8999s | ~1.6 |

I was looking at making `TabulatedFluidProperties` faster, and in combination with #10996, the time to sample a value and two first derivatives is now over twice as fast (numbers for 1 million samples):

| Before | After | Speedup |
|--|--|--|
| 15.1424s | 6.57069s | ~2.2 |

This change will also make objects that use bicubic spline interpolation faster, such as `BicubicSplineFunction`.

Closes #10999